requirements.txt
uvicorn
fastapi
requests
pytest
pyvespa
docker
streamlit
aiofiles
langchain_community==0.2.1
langchain_fireworks==0.1.3
langchain==0.2.1
python-docx==1.1.2
docx2txt==0.8
json-repair==0.20.1
pandas==2.2.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
unidecode==1.3.8
aiobotocore==2.13.1
langfuse==2.44.0
google-generativeai==0.8.3
markitdown==0.0.1a3
pypdf==5.6.0
Pillow==10.2.0
psycopg2-binary
agno
google-genai
faker
sqlalchemy
psycopg[binary]